+++ src/commands/convert.ts
@@ -3,6 +3,12 @@ import { latexToPretext } from "../latextopretext";
 import { markdownToPretext } from "md2ptx";
 import { pretextOutputChannel } from "../ui";
 import { convertToPretext } from "../importFiles";
+import { processLatexViaUnified } from "@unified-latex/unified-latex";
+import {
+  unifiedLatexToPretext,
+  xmlCompilePlugin,
+} from "@unified-latex/unified-latex-to-pretext";
+import { formatPTX } from "../formatter";
 
 export function cmdConvertToPretext() {
   console.log("Converting to PreTeXt");
@@ -29,15 +35,45 @@ export function cmdLatexToPretext() {
       ? editor.document.lineAt(selection.start.line).range
       : new Range(selection.start, selection.end);
 
-    var initialText = editor.document.getText(selectionRange);
+    //console.log("selectionRange is", selectionRange);
 
-    var newText = latexToPretext(initialText);
+    const initialText = editor.document.getText(selectionRange);
 
+    //var newText = latexToPretext(initialText);
+    let newText = convertWithUnified(initialText);
+
+    // Remove the starting <p> tag if we selected text in the middle of a line.
+    const pTagMatch = newText.match(/^<p>/);
+    if (pTagMatch && selectionRange.start.character > 0) {
+      newText = newText.replace(/^<p>/, "");
+    }
+
+    // Format the new text.
+    const formattedNewText = formatPTX(newText);
     editor.edit((editbuilder) => {
-      editbuilder.replace(selectionRange, newText);
+      editbuilder.replace(selectionRange, formattedNewText);
     });
   }
 }
+
+  
+function convertWithUnified(text: string) {
+  const convert = (value: string) =>
+    processLatexViaUnified()
+      .use(unifiedLatexToPretext, { producePretextFragment: true })
+      .use(xmlCompilePlugin)
+      .processSync({ value });
+
+  pretextOutputChannel.append("Converting selected text to PreTeXt.\n");
+  if (convert(text).messages) {
+    for (let message of convert(text).messages) {
+      pretextOutputChannel.appendLine(message.message);
+      console.log(message);
+    }
+  }
+  return convert(text).value as string;
+}
+
 
 export function cmdMarkdownToPretext() {
   const editor = window.activeTextEditor;
@@ -55,4 +91,5 @@ export function cmdMarkdownToPretext() {
       editbuilder.replace(selectionRange, newText);
     });
   }
-}
+
+