+++ src/commands/runPtx.ts
@@ -10,7 +10,7 @@ export async function runPretext(
   ptxExec: string,
   ptxCommand: string,
   ptxOptions: string,
-  passedPath: string = "",
+  passedPath: string = ""
 ): Promise<void> {
   return window.withProgress(
     {
@@ -24,7 +24,7 @@ export async function runPretext(
         var progressUpdate = "Starting up...";
         const interval = setInterval(
           () => progress.report({ message: progressUpdate }),
-          1000,
+          1000
         );
         let fullCommand = cli.cmd() + " " + ptxCommand + " " + ptxOptions;
         let status = "ready"; //for statusbaritem
@@ -35,7 +35,7 @@ export async function runPretext(
           let capturedErrors: string[] = [];
           pretextOutputChannel.clear();
           pretextOutputChannel.appendLine(
-            "\n\nNow running `" + fullCommand + "`...",
+            "\n\nNow running `" + fullCommand + "`..."
           );
           progressUpdate = "Running " + fullCommand;
           var ptxRun = spawn(fullCommand, [], {
@@ -50,7 +50,7 @@ export async function runPretext(
               data
                 .toString()
                 .includes(
-                  "Your built project will soon be available to the public at:",
+                  "Your built project will soon be available to the public at:"
                 )
             ) {
               capturedOutput.push(data);
@@ -63,7 +63,7 @@ export async function runPretext(
             ) {
               pretextOutputChannel.appendLine(`${data}`);
               pretextOutputChannel.append(
-                "(this local server will remain running until you close vs-code)\n",
+                "(this local server will remain running until you close vs-code)\n"
               );
               capturedOutput.push(data);
               console.log("Using view. Status should change back");
@@ -84,7 +84,7 @@ export async function runPretext(
             console.log(code);
             if (ptxRun.killed) {
               pretextOutputChannel.appendLine(
-                "...PreTeXt command terminated early.",
+                "...PreTeXt command terminated early."
               );
               console.log("Process killed");
             } else {
@@ -100,7 +100,7 @@ export async function runPretext(
                 .showErrorMessage(
                   "PreTeXt encountered one or more errors",
                   "Show Log",
-                  "Dismiss",
+                  "Dismiss"
                 )
                 .then((option) => {
                   if (option === "Show Log") {
@@ -110,14 +110,14 @@ export async function runPretext(
             } else {
               console.log(
                 "PreTeXt command finished successfully; code =",
-                code,
+                code
               );
               if (ptxCommand === "build") {
                 window
                   .showInformationMessage(
                     "Build successful! You can preview your output now.",
                     "Dismiss",
-                    "View log",
+                    "View log"
                   )
                   .then((option) => {
                     if (option === "View log") {
@@ -132,7 +132,7 @@ export async function runPretext(
                       .splice(-1)[0]
                       .split("soon be available to the public at:")
                       .slice(-1)[0]
-                      .trim(),
+                      .trim()
                   );
                   console.log("Opening site at: ", siteURL);
                   window
@@ -140,7 +140,7 @@ export async function runPretext(
                       "Deploy successful! You can view your deployed site now.",
                       "Visit site",
                       "View log",
-                      "Dismiss",
+                      "Dismiss"
                     )
                     .then((option) => {
                       if (option === "Visit site") {
@@ -163,7 +163,7 @@ export async function runPretext(
           });
         }
       });
-    },
+    }
   );
 }
 
@@ -171,11 +171,11 @@ export async function runThenOpen(
   ptxExec: string,
   ptxCommand: string,
   ptxOptions: string,
-  folder: string,
+  folder: string
 ) {
   await runPretext(ptxExec, ptxCommand, ptxOptions, folder);
   commands.executeCommand(
     "vscode.openFolder",
-    path.join(folder, "new_pretext_project"),
+    path.join(folder, "new_pretext_project")
   );
 }