+++ src/extension.ts
@@ -7,7 +7,7 @@ import {
   TextEdit,
   commands,
 } from "vscode";
-import { formatPTX } from "./formatter";
+import { formatPretextDocument } from "./formatter";
 import * as utils from "./utils";
 
 import {
@@ -109,9 +109,9 @@ export async function activate(context: ExtensionContext) {
       "pretext",
       {
         provideDocumentFormattingEdits(document: TextDocument): TextEdit[] {
-          return formatPTX(document);
+          return formatPretextDocument(document);
         },
-      },
+      }
     );
 
     context.subscriptions.push(formatter);
@@ -125,7 +125,7 @@ export async function activate(context: ExtensionContext) {
     //commands.registerCommand("pretext-tools.experiment", utils.experiment),
     commands.registerCommand(
       "pretext-tools.selectPretextCommand",
-      cmdSelectCommand,
+      cmdSelectCommand
     ),
     commands.registerCommand("pretext-tools.buildAny", cmdBuildAny),
     commands.registerCommand("pretext-tools.buildLast", cmdBuildLast),
@@ -139,7 +139,7 @@ export async function activate(context: ExtensionContext) {
     commands.registerCommand("pretext-tools.latexToPretext", cmdLatexToPretext),
     commands.registerCommand(
       "pretext-tools.convertToPretext",
-      cmdConvertToPretext,
+      cmdConvertToPretext
     ),
     commands.registerCommand(
       "pretext-tools.markdownToPretext",
@@ -159,15 +159,15 @@ export async function activate(context: ExtensionContext) {
         "Targets are now:" +
           targetSelection.map(function (obj: Target) {
             return " " + obj.label;
-          }),
+          })
       );
       window.showInformationMessage(
         "Refreshed list of targets.  Targets are now:" +
           targetSelection.map(function (obj: Target) {
             return " " + obj.label;
-          }),
+          })
       );
-    }),
+    })
   );
 
   // Start the LSP
@@ -176,12 +176,12 @@ export async function activate(context: ExtensionContext) {
   } catch {
     console.log("Error starting LSP client");
     pretextOutputChannel.appendLine(
-      "Error starting language server.  Some features may not be available.",
+      "Error starting language server.  Some features may not be available."
     );
   }
 
   pretextOutputChannel.appendLine(
-    "PreTeXt related commands are available through the PreTeXt status bar menu or the command pallet (CTRL+SHIFT+P).",
+    "PreTeXt related commands are available through the PreTeXt status bar menu or the command pallet (CTRL+SHIFT+P)."
   );
 }