+++ src/lsp-server/completions/get-completions.ts
@@ -35,7 +35,7 @@ const completionCache: CompletionItem[] = [];
 
 async function getCompletionType(
   doc: TextDocument,
-  pos: Position,
+  pos: Position
 ): Promise<CompletionType> {
   // For now we assume that open tags are on the same line as the cursor.  TODO: fix this.
   const linePrefix = doc.getText(lineToPosition(pos));
@@ -56,7 +56,7 @@ async function getCompletionType(
 }
 
 export async function getCompletions(
-  params: TextDocumentPositionParams,
+  params: TextDocumentPositionParams
 ): Promise<CompletionItem[] | null> {
   const uri = params.textDocument.uri;
   const info = getDocumentInfo(uri);
@@ -248,7 +248,7 @@ export async function getCompletions(
 function getExtraCompletions(
   element: string,
   range: Range,
-  schema: Schema,
+  schema: Schema
 ): CompletionItem[] {
   let extraCompletions: CompletionItem[] = [];
   for (let [key, item] of Object.entries(EXTRA_ELEMENT_SNIPPETS)) {
@@ -328,7 +328,7 @@ function getExtraCompletions(
  * Retrieve the full `item` details for the abbreviated completion item.
  */
 export async function getCompletionDetails(
-  item: CompletionItem,
+  item: CompletionItem
 ): Promise<CompletionItem> {
   return completionCache[item.data];
 }